+++ src/Controller/ArtistController.php
@@ -2,53 +2,68 @@
 
 namespace App\Controller;
 
-use App\Service\CreateUpdateDeleteServiceInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use App\Validator\ArtistValidateInterface;
 
-class ArtistController extends AbstractController
+class ArtistController extends BaseController
 {
-    private $CUDService;
-
-    public function __construct(CreateUpdateDeleteServiceInterface $CUDService)
-    {
-        $this->CUDService = $CUDService;
-    }
-
     /**
      * @Route("/createArtist", name="createArtist")
      * @param Request $request
+     * @return
      */
-    public function create(Request $request)
+    public function create(Request $request, ArtistValidateInterface $artistValidate)
     {
-        //ToDo Call Validator
+        //Validation
+        $validateResult = $artistValidate->artistValidator($request, 'create');
+        if (!empty($validateResult))
+        {
+            $resultResponse = new Response($validateResult, Response::HTTP_OK, ['content-type' => 'application/json']);
+            $resultResponse->headers->set('Access-Control-Allow-Origin', '*');
+            return $resultResponse;
+        }
+        //
 
         $result = $this->CUDService->create($request, "Artist");
-        return $result;
+        return $this->response($result, self::CREATE);
     }
 
     /**
      * @Route("/updateArtist", name="updateArtist")
      * @param Request $request
+     * @return
      */
-    public function update(Request $request)
+    public function update(Request $request, ArtistValidateInterface $artistValidate)
     {
-        //ToDo Call Validator
-
+        $validateResult = $artistValidate->artistValidator($request, 'update');
+        if (!empty($validateResult))
+        {
+            $resultResponse = new Response($validateResult, Response::HTTP_OK, ['content-type' => 'application/json']);
+            $resultResponse->headers->set('Access-Control-Allow-Origin', '*');
+            return $resultResponse;
+        }
         $result = $this->CUDService->update($request, "Artist");
-        return $result;
+        return $this->response($result, self::UPDATE);
     }
 
     /**
      * @Route("/deleteArtist", name="deleteArtist")
      * @param Request $request
+     * @return
      */
-    public function delete(Request $request)
+    public function delete(Request $request, ArtistValidateInterface $artistValidate)
     {
-        //ToDo Call Validator
-
+        $validateResult = $artistValidate->artistValidator($request, 'delete');
+        if (!empty($validateResult))
+        {
+            $resultResponse = new Response($validateResult, Response::HTTP_OK, ['content-type' => 'application/json']);
+            $resultResponse->headers->set('Access-Control-Allow-Origin', '*');
+            return $resultResponse;
+        }
         $result = $this->CUDService->delete($request, "Artist");
-        return $result;
+        return $this->response($result, self::DELETE);
+
     }
 }
