+++ src/Controller/AuctionPaintingController.php
@@ -3,54 +3,70 @@
 namespace App\Controller;
 
 use App\Service\CreateUpdateDeleteServiceInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use App\Validator\AuctionPaintingValidate;
+use App\Validator\AuctionPaintingValidateInterface;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-
-class AuctionPaintingController extends AbstractController
-
+class AuctionPaintingController extends BaseController
 {
-    private $CUDService;
-
-    public function __construct(CreateUpdateDeleteServiceInterface $CUDService)
-    {
-        $this->CUDService = $CUDService;
-    }
-
     /**
      * @Route("/createAuctionPainting", name="createAuctionPainting")
      * @param Request $request
+     * @return
      */
-    public function create(Request $request)
+    public function create(Request $request, AuctionPaintingValidateInterface $auctionPaintingValidate)
     {
-        //ToDo Call Validator
+        //Validation
+        $validateResult = $auctionPaintingValidate->auctionPaintingValidator($request, 'create');
+        if (!empty($validateResult))
+        {
+            $resultResponse = new Response($validateResult, Response::HTTP_OK, ['content-type' => 'application/json']);
+            $resultResponse->headers->set('Access-Control-Allow-Origin', '*');
+            return $resultResponse;
+        }
+        //
 
         $result = $this->CUDService->create($request, "AuctionPainting");
-        return $result;
+        return $this->response($result, self::CREATE);
     }
 
     /**
      * @Route("/updateAuctionPainting", name="updateAuctionPainting")
      * @param Request $request
+     * @return
      */
-    public function update(Request $request)
+    public function update(Request $request, AuctionPaintingValidateInterface $auctionPaintingValidate)
     {
-        //ToDo Call Validator
-
+        $validateResult = $auctionPaintingValidate->auctionPaintingValidator($request, 'update');
+        if (!empty($validateResult))
+        {
+            $resultResponse = new Response($validateResult, Response::HTTP_OK, ['content-type' => 'application/json']);
+            $resultResponse->headers->set('Access-Control-Allow-Origin', '*');
+            return $resultResponse;
+        }
         $result = $this->CUDService->update($request, "AuctionPainting");
-        return $result;
+        return $this->response($result, self::UPDATE);
     }
 
     /**
      * @Route("/deleteAuctionPainting", name="deleteAuctionPainting")
      * @param Request $request
+     * @return
      */
-    public function delete(Request $request)
+    public function delete(Request $request, AuctionPaintingValidateInterface $auctionPaintingValidate)
     {
-        //ToDo Call Validator
-
+        $validateResult = $auctionPaintingValidate->auctionPaintingValidator($request, 'delete');
+        if (!empty($validateResult))
+        {
+            $resultResponse = new Response($validateResult, Response::HTTP_OK, ['content-type' => 'application/json']);
+            $resultResponse->headers->set('Access-Control-Allow-Origin', '*');
+            return $resultResponse;
+        }
         $result = $this->CUDService->delete($request, "AuctionPainting");
-        return $result;
+        return $this->response($result, self::DELETE);
+
     }
 }
